# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "A service that permits to handle converted WISP requests for technical support",
    "termsOfService" : "https://www.pagopa.gov.it/",
    "title" : "wisp-converter-technical-support",
    "version" : "0.2.1-7-feat-flow-analysis-enhancements"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
    "/enhanced/monitoring/receipts/pending" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Retrieve the list of pending receipt that could have not been sent to creditor institution",
        "operationId" : "extractPendingReceipts",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PendingReceiptsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReceiptsStatusSnapshotResponse"
                }
              }
            },
            "description" : "Successfully extracted data",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get pending receipts",
        "tags" : [ "Technical Support - Enhanced features" ]
      }
    },
    "/enhanced/monitoring/receipts/snapshot" : {
      "get" : {
        "description" : "Retrieve a snapshot of receipt status at specific slot time.",
        "operationId" : "extractReceiptSnapshot",
        "parameters" : [ {
          "description" : "Lower limit date time (in yyyy-MM-ddThh:mm:ss) in 'Europe/Rome' timezone",
          "example" : "2024-01-01T12:00:00",
          "in" : "query",
          "name" : "dateFrom",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Lower limit date time (in yyyy-MM-ddThh:mm:ss) in 'Europe/Rome' timezone",
            "example" : "2024-01-01T12:00:00"
          }
        }, {
          "description" : "Upper limit date time (in yyyy-MM-ddThh:mm:ss) in 'Europe/Rome' timezone",
          "example" : "2024-01-01T13:00:00",
          "in" : "query",
          "name" : "dateTo",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Upper limit date time (in yyyy-MM-ddThh:mm:ss) in 'Europe/Rome' timezone",
            "example" : "2024-01-01T13:00:00"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReceiptsStatusSnapshotResponse"
                }
              }
            },
            "description" : "Successfully extracted snapshot",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get snapshot of receipt status",
        "tags" : [ "Technical Support - Enhanced features" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/enhanced/organizations/{organizationId}/iuv/{iuv}/details" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Retrieve the details of a payment searching by IUV, analyzing all its status and events.",
        "operationId" : "findByIuv",
        "parameters" : [ {
          "description" : "The identifier of the creditor institution",
          "example" : 12345678900,
          "in" : "path",
          "name" : "organizationId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "The identifier of the creditor institution",
            "example" : 12345678900
          }
        }, {
          "description" : "The IUV code of the payment",
          "example" : "000000999900000",
          "in" : "path",
          "name" : "iuv",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "The IUV code of the payment",
            "example" : "000000999900000"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string",
                "description" : "The filters to be applied on the request"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentFlowsResponse"
                }
              }
            },
            "description" : "Successfully retrieved details",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get payment details searching by IUV",
        "tags" : [ "Technical Support - Enhanced features" ]
      }
    },
    "/enhanced/organizations/{organizationId}/iuv/{iuv}/status" : {
      "get" : {
        "description" : "Retrieve the status of a payment, analyzing also the status fo the receipt.",
        "operationId" : "findStatusByIuv",
        "parameters" : [ {
          "description" : "Creditor institution identifier",
          "example" : 12345678900,
          "in" : "path",
          "name" : "organizationId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Creditor institution identifier",
            "example" : 12345678900
          }
        }, {
          "description" : "IUV payment code",
          "example" : 982273480765907,
          "in" : "path",
          "name" : "iuv",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "IUV payment code",
            "example" : 982273480765907
          }
        }, {
          "description" : "Lower limit date (in yyyy-MM-dd)",
          "example" : "2024-01-01",
          "in" : "query",
          "name" : "dateFrom",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Lower limit date (in yyyy-MM-dd)",
            "example" : "2024-01-01"
          }
        }, {
          "description" : "Upper limit date (in yyyy-MM-dd)",
          "example" : "2024-01-01",
          "in" : "query",
          "name" : "dateTo",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Upper limit date (in yyyy-MM-dd)",
            "example" : "2024-01-01"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentFlowStatusResponse"
                }
              }
            },
            "description" : "Successfully retrieved payment status",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get payment status searching by IUV",
        "tags" : [ "Technical Support - Enhanced features" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/enhanced/organizations/{organizationId}/notice-number/{noticeNumber}/details" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Retrieve the details of a payment searching by notice number, analyzing all its status and events.",
        "operationId" : "findByNoticeNumber",
        "parameters" : [ {
          "description" : "Creditor institution identifier",
          "example" : 12345678900,
          "in" : "path",
          "name" : "organizationId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Creditor institution identifier",
            "example" : 12345678900
          }
        }, {
          "description" : "Notice number payment code",
          "example" : 982273480765907,
          "in" : "path",
          "name" : "noticeNumber",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Notice number payment code",
            "example" : 982273480765907
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string",
                "description" : "The filters to be applied on the request"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentFlowsResponse"
                }
              }
            },
            "description" : "Successfully retrieved details",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get payment details searching by notice number",
        "tags" : [ "Technical Support - Enhanced features" ]
      }
    },
    "/enhanced/session-id/{sessionId}/details" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Retrieve the details of a payment searching by session ID, analyzing all its status and events.",
        "operationId" : "findBySessionId",
        "parameters" : [ {
          "description" : "The session ID related to the single payment flow",
          "example" : "12345678900_aa0518a9-194d-4de2-999f-e462c3746e24",
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "The session ID related to the single payment flow",
            "example" : "12345678900_aa0518a9-194d-4de2-999f-e462c3746e24"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string",
                "description" : "The filters to be applied on the request"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentFlowsResponse"
                }
              }
            },
            "description" : "Successfully retrieved event",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get payment details searching by session ID",
        "tags" : [ "Technical Support - Enhanced features" ]
      }
    },
...
    "/organizations/{organizationId}/iuv/{iuv}" : {
      "get" : {
        "operationId" : "findByIuv_1",
        "parameters" : [ {
          "description" : "Creditor institution identifier",
          "example" : 12345678900,
          "in" : "path",
          "name" : "organizationId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Creditor institution identifier",
            "example" : 12345678900
          }
        }, {
          "description" : "IUV payment code",
          "example" : 982273480765907,
          "in" : "path",
          "name" : "iuv",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "IUV payment code",
            "example" : 982273480765907
          }
        }, {
          "description" : "Lower limit date",
          "example" : "2024-03-18",
          "in" : "query",
          "name" : "dateFrom",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Lower limit date",
            "example" : "2024-03-18"
          }
        }, {
          "description" : "Upper limit date",
          "example" : "2024-03-18",
          "in" : "query",
          "name" : "dateTo",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Upper limit date",
            "example" : "2024-03-18"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReEventResponse"
                }
              }
            },
            "description" : "Successfully retrieved event",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "tags" : [ "Technical Support" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/organizations/{organizationId}/notice-number/{noticeNumber}" : {
      "get" : {
        "operationId" : "findByNoticeNumber_1",
        "parameters" : [ {
          "description" : "Creditor institution identifier",
          "example" : 12345678900,
          "in" : "path",
          "name" : "organizationId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Creditor institution identifier",
            "example" : 12345678900
          }
        }, {
          "description" : "NAV payment code",
          "example" : 3982273480765907,
          "in" : "path",
          "name" : "noticeNumber",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "NAV payment code",
            "example" : 3982273480765907
          }
        }, {
          "description" : "Lower limit date",
          "example" : "2024-03-18",
          "in" : "query",
          "name" : "dateFrom",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Lower limit date",
            "example" : "2024-03-18"
          }
        }, {
          "description" : "Upper limit date",
          "example" : "2024-03-18",
          "in" : "query",
          "name" : "dateTo",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Upper limit date",
            "example" : "2024-03-18"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReEventResponse"
                }
              }
            },
            "description" : "Successfully retrieved event",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "tags" : [ "Technical Support" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/session-id/{sessionId}" : {
      "get" : {
        "operationId" : "findBySessionId_1",
        "parameters" : [ {
          "description" : "Session identifier of the flow",
          "example" : "12345678900_aa0518a9-194d-4de2-999f-e462c3746e24",
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Session identifier of the flow",
            "example" : "12345678900_aa0518a9-194d-4de2-999f-e462c3746e24"
          }
        }, {
          "description" : "Data di ricerca DA",
          "example" : "2024-03-18",
          "in" : "query",
          "name" : "dateFrom",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Data di ricerca DA",
            "example" : "2024-03-18"
          }
        }, {
          "description" : "Data di ricerca A",
          "example" : "2024-03-18",
          "in" : "query",
          "name" : "dateTo",
          "required" : true,
          "schema" : {
            "type" : "string",
            "description" : "Data di ricerca A",
            "example" : "2024-03-18"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReEventResponse"
                }
              }
            },
            "description" : "Successfully retrieved event",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "tags" : [ "Technical Support" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    }
  },
  "components" : {
    "schemas" : {
...
      "PaymentFlowStatusResponse" : {
        "type" : "object",
        "properties" : {
          "lower_bound_date" : {
            "type" : "string",
            "format" : "date"
          },
          "payments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PaymentFlowStatus"
            }
          },
          "upper_bound_date" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
...
      "PaymentFlowsResponse" : {
        "type" : "object",
        "properties" : {
          "flows" : {
            "type" : "array",
            "description" : "List of flows executed on certain payment position",
            "items" : {
              "$ref" : "#/components/schemas/PaymentFlow"
            }
          },
          "metadata" : {
            "$ref" : "#/components/schemas/PaymentFlowMetadata"
          }
        }
      },
...
      "PendingReceiptsRequest" : {
        "required" : [ "creditor_institution", "lower_bound_date", "upper_bound_date" ],
        "type" : "object",
        "properties" : {
          "creditor_institution" : {
            "type" : "string",
            "description" : "The identifier of the creditor institution",
            "example" : "12345678901"
          },
          "iuvs" : {
            "uniqueItems" : true,
            "type" : "array",
            "description" : "The list of IUV payment codes to be used as a search filter",
            "items" : {
              "type" : "string",
              "description" : "The list of IUV payment codes to be used as a search filter"
            }
          },
          "lower_bound_date" : {
            "type" : "string",
            "description" : "Lower limit date time (in yyyy-MM-ddThh:mm:ss) in 'Europe/Rome' timezone",
            "format" : "date-time"
          },
          "upper_bound_date" : {
            "type" : "string",
            "description" : "Upper limit date time (in yyyy-MM-ddThh:mm:ss) in 'Europe/Rome' timezone",
            "format" : "date-time"
          }
        }
      },
...
      "ReEventResponse" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer",
            "description" : "Totale dei record trovati",
            "format" : "int32",
            "example" : 100
          },
          "data" : {
            "type" : "array",
            "description" : "Lista dei record trovati",
            "items" : {
              "$ref" : "#/components/schemas/ReEvent"
            }
          }
        }
      },
      "ReceiptsStatusSnapshot" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
      "ReceiptsStatusSnapshotResponse" : {
        "type" : "object",
        "properties" : {
          "lower_bound_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "snapshot" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReceiptsStatusSnapshot"
            }
          },
          "upper_bound_date" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      }
    },
    "securitySchemes" : {
      "ApiKey" : {
        "description" : "The API key to access this function app.",
        "in" : "header",
        "name" : "Ocp-Apim-Subscription-Key",
        "type" : "apiKey"
      }
    }
  }
}